[#28] Fix bugs in location filter
This changes adds support for multiple-word filters (e.g. United States), and
also adds the filters as an AND into the query (and not OR, or whatever is the
configured default in Elastic Search). It means that the filter has to be true,
regardless of the rest of the query.