[#28] Fix bug when querying multiple words and filters
For example, say that we're querying for "heart disease" in "Israel". Before
this patch, the ElasticSearch query would be:

    heart disease AND location:Israel

Adding parenthesis and the default connectors to make the error clearer, it
would become:

    heart OR (disease AND location:Israel)

Which isn't what we want. With this patch, the query string is put inside
parenthesis, as:

    (heart disease) AND location:Israel

So it works as expected.