[#28] Improving how we mock the API in the tests
There were two changes. First, instead of mocking the API with the query
parameters, as "/search?q=foo&page=2", we use nock's `query`, so the
order of the query parameters doesn't matter ("/search?page=2&q=foo" would
match the same mock).

The second change was creating a `mockApiResponses` helper. A request to
"/search" in this codebase creates requests for two API endpoints (so far,
it'll grow). So they both need to be mocked every time, even if the test in
question only cares for one of them. Encapsulating this makes it easier to
change the tests. For instance, we won't have to fix every test when the API
endpoint changes.