[#28] Adds entire query into context, instead of just filters
We'll have to modify the filters in the query in some time. For example, a
filter on `registration_date` range would neeed to be "[2012-01-01 TO
2013-01-01]". We also need the actual `registration_date` values to add in the
form, so we need to separate filters from what's in the query.

I thought about encapsulating this logic of converting query values into
elasticsearch filters inside the `agents/trials.js`, but that felt like it would
add too much coupling. Maybe this is an idea we should revisit in the future,
when we see how things work out.