[#55] Add basic end to end smoke test
It open our index page, clicks on the search button, clicks on the first
result's title and tries to download it as a JSON.

To avoid always running our smoke tests, I had to reorganize how we run Mocha.
Our e2e tests MUST have "e2e" in the topmost "describe" block. There's a pull
request in Mocha to include a `--tags` option which would simplify this pattern
in the future.