Reload server only once when "rev" file changes
We watch for changes in our SCSS files, and then run the "rev" task. This task
creates the stylesheets and JS files in `dist/styles` and `dist/js`, and then
the rev manifest in `dist/rev-manifest.json`. If we don't ignore `dist/styles`
and `dist/js`, `nodemon` will reload for each new file in these directories. We
can't ignore the `dist/` folder altogether because we need to reload when
`dist/rev-manifest.json` changes.