[#170] Escape queries to ElasticSearch
As we're using ES' query strings, we need to escape the values we send through
to make sure we don't end up sending some invalid queries (like "foo("). Most
of the sanitization code comes from
https://github.com/lanetix/node-elasticsearch-sanitize/. I couldn't use their
package directly because they also escape whitespaces and other things (e.g.
OR, AND, NOT), which we don't want. The code is trivial, though, so no big
deal.