[#209] Fix bugs with published results filter
If `has_published_results == false`, we weren't adding any filters, so we would
get all trials (including those with published results). This is because we
relied on simply testing for truthyness, and not for the actual true/false
value. As `undefined` is falsy in JavaScript, there was no difference between
setting `has_published_results = false` and not setting `has_published_results`
at all.

I've also fixed the bug where the study results filter wasn't persistent. The
problem here is that we were testing for "true" and "false" strings, and not
boolean values. This was the case before we added Joi for validation. As Joi
also converts the values, it stopped working.