[#198] Add data categories and allow links to be contributed
The link allow users to contribute for example blog posts or trials webpages.
Ideally, when a user sends a contribution, she'll either upload a file or add a
link (or both). However, I've chosen not to require this (for now, at least)
because of the added complexity to the form. It's easy to ignore/remove bogus
contributions as well, and we reduce friction by being more lenient with what
we receive.

For the categories, I've chosen to create another entity and save those to the
database so it's easier to tweak names and add translations in the future. This
has the drawback of the added complexity, a small performance hit, and making
it harder for us to accept free-form categories if needed later. However, I
thought it was worth the hassle.