[#205] Don't show advanced filters before loading JS
We did show the filters before loading JS to allow people that have JS disabled
to see them (we use JS to toggle the filters). However, regular users were
seeing a flash of the advanced filters before the browser had a change to load
JS and hide it.

Now, instead of relying on Modernizr to fallback, I've added an
`advanced_search` query value in the search pages. If the user doesn't have JS
enabled, the "Advanced search" toggle will redurect her to the same page with
`?advanced_search=true`, which shows the advanced search options. If she has JS
enabled, we prevent the link behaviour and show the filters using JS.