[#501] Refactor server to use multiple connections
This allow us to separate the FDA-related pages from the rest by binding it to
a different port and writing it inside a plugin. This is an initial step in a
somewhat large refactoring to extract functionalities into plugins to make them
more isolated and, hopefully, easier to control their complexity.

See
https://medium.com/@dstevensio/manifests-plugins-and-schemas-organizing-your-hapi-application-68cf316730ef#.un78wyp7f
for more information on the long-term goal.

opentrials/opentrials#501